Networking

+++ Assets/Events/WorkEvent.cs
@@ -9,10 +9,10 @@ namespace MarsTS.Events {
     public class WorkEvent : AbstractEvent {
 
 		public ISelectable Unit { get; private set; }
-		public float WorkRequired { get; private set; }
-		public float CurrentWork { get; private set; }
+		public int WorkRequired { get; private set; }
+		public int CurrentWork { get; private set; }
 
-		public WorkEvent (EventAgent _source, ISelectable _unit, float _workRequired, float _currentWork) : base("work", _source) {
+		public WorkEvent (EventAgent _source, ISelectable _unit, int _workRequired, int _currentWork) : base("work", _source) {
 			Unit = _unit;
 			WorkRequired = _workRequired;
 			CurrentWork = _currentWork;
